requirements.txt
networkx
torch
torch-geometric
pyyaml